Combine source controllers
